Knife and knuckles removed from the game (they were responsible for at least two stability bugs - with loading graphics and throwing). Anyway, a better system for adding new items is needed.
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@345 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/multiplay.cpp
@@ -1195,7 +1195,7 @@ int Net::recv_unit_data_size()
 	} else { //end send
 		//info->printstr("pd recv end\n");
 		pd_remote->size = size;
-		g_console->printf("pd recv end. size %d", pd_remote->size);
+		g_console->printf("pd recv end. size %d", (int)pd_remote->size);
 		remote.SEND = 1;
 	}
 	return 1;
